# docs/swagger.yaml
definitions:
...
  dto.CreateRole:
    properties:
      actions:
        items:
          $ref: '#/definitions/dto.MenusAction'
        type: array
      active:
        type: boolean
      name:
        type: string
      object_actions:
        items:
          properties:
            id:
              type: integer
            name:
              type: string
            parentId:
              type: integer
            path:
              type: string
            state:
              type: integer
            type:
              type: string
          required:
          - id
          - name
          - path
          - type
          type: object
        type: array
    required:
    - actions
    - name
    - object_actions
    type: object
...
  dto.DeleteRoleBulk:
    properties:
      role_ids:
        items:
          type: integer
        minItems: 1
        type: array
    required:
    - role_ids
    type: object
...
  dto.MenusAction:
    properties:
      apiid:
        items:
          type: integer
        type: array
      menuId:
        type: integer
    required:
    - menuId
    type: object
...
  dto.SetAccessRole:
    properties:
      actions:
        items:
          $ref: '#/definitions/dto.MenusAction'
        type: array
      roleId:
        type: integer
    required:
    - actions
    - roleId
    type: object
...
  middlewares.ResponseError:
    properties:
      message: {}
      request_id:
        type: string
      time:
        type: string
    type: object
  models.Action:
    properties:
      api:
        type: string
      created_at:
        type: string
      default:
        type: boolean
      id:
        type: integer
      method:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
  models.ActionsMenus:
    properties:
      actionId:
        type: integer
      actions:
        $ref: '#/definitions/models.Action'
      created_at:
        type: string
      id:
        type: integer
      menuId:
        type: integer
      updated_at:
        type: string
    type: object
...
  models.Menu:
    properties:
      actions_menus:
        items:
          $ref: '#/definitions/models.ActionsMenus'
        type: array
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      parentId:
        type: integer
      path:
        type: string
      updated_at:
        type: string
    type: object
  models.Paginate:
    properties:
      page:
        type: string
      page_count:
        type: integer
      page_size:
        type: string
      total:
        type: integer
    type: object
...
  models.Role:
    properties:
      created_at:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      name:
        type: string
      object_menus:
        items: {}
        type: array
      roles_menus:
        items:
          $ref: '#/definitions/models.RolesMenus'
        type: array
      updated_at:
        type: string
    type: object
  models.RolesMenus:
    properties:
      created_at:
        type: string
      id:
        type: integer
      menu:
        $ref: '#/definitions/models.Menu'
      menu_id:
        type: integer
      role:
        $ref: '#/definitions/models.Role'
      roleId:
        type: integer
      updated_at:
        type: string
    type: object
...
  utils.Response:
    properties:
      data: {}
      message: {}
      statusCode:
        type: integer
    type: object
  utils.ResponsePaginate:
    properties:
      data: {}
      key:
        type: string
      message: {}
      meta:
        $ref: '#/definitions/models.Paginate'
      statusCode:
        type: integer
    type: object
info:
  contact:
    name: PT Tata Kelola Dokumen
  description: Aplikasi Assessment & Monitoring
  license:
    name: Private License
  title: EJM Assessment & Monitoring API Documentation
  version: "1.0"
paths:
...
  /roles:
    get:
      consumes:
      - application/json
      parameters:
      - description: Halaman
        in: query
        name: page
        required: true
        type: string
      - description: Jumlah Data Per halaman
        in: query
        name: page_size
        required: true
        type: string
      - description: Mencari Data
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ResponsePaginate'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/middlewares.ResponseError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middlewares.ResponseError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/middlewares.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/middlewares.ResponseError'
      summary: API untuk Ambil daftar Role
      tags:
      - Role
    post:
      consumes:
      - application/json
      parameters:
      - description: Daftar Role Baru
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/dto.CreateRole'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.Role'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/middlewares.ResponseError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middlewares.ResponseError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/middlewares.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/middlewares.ResponseError'
      summary: API untuk Daftar Role Baru
      tags:
      - Role
  /roles/{id}:
    get:
      consumes:
      - application/json
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/middlewares.ResponseError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middlewares.ResponseError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/middlewares.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/middlewares.ResponseError'
      summary: API untuk Get Role Berdasarkan ID
      tags:
      - Role
    patch:
      consumes:
      - application/json
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: string
      - description: Update Role
        in: body
        name: roles
        required: true
        schema:
          $ref: '#/definitions/dto.CreateRole'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/middlewares.ResponseError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middlewares.ResponseError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/middlewares.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/middlewares.ResponseError'
      summary: API untuk Update Role Berdasarkan ID
      tags:
      - Role
  /roles/delete-access-role:
    delete:
      consumes:
      - application/json
      parameters:
      - description: role id
        in: query
        name: roleId
        required: true
        type: string
      - description: menu id
        in: query
        name: menuId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.ResponsePaginate'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/middlewares.ResponseError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middlewares.ResponseError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/middlewares.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/middlewares.ResponseError'
      summary: API Delete Akses Role
      tags:
      - Role
  /roles/delete-bulk:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Hapus Role banyak
        in: body
        name: roleIds
        required: true
        schema:
          $ref: '#/definitions/dto.DeleteRoleBulk'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/middlewares.ResponseError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middlewares.ResponseError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/middlewares.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/middlewares.ResponseError'
      summary: API untuk Hapus Role banyak
      tags:
      - Role
  /roles/delete/{roleId}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Role Id
        in: path
        name: roleId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/middlewares.ResponseError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middlewares.ResponseError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/middlewares.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/middlewares.ResponseError'
      summary: API untuk Delete Role Berdasarkan ID
      tags:
      - Role
  /roles/set-access-role:
    post:
      consumes:
      - application/json
      parameters:
      - description: Akses Role
        in: body
        name: login
        required: true
        schema:
          $ref: '#/definitions/dto.SetAccessRole'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/middlewares.ResponseError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middlewares.ResponseError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/middlewares.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/middlewares.ResponseError'
      summary: Set Akses Role
      tags:
      - Auth
  /roles/update-access-role:
    patch:
      consumes:
      - application/json
      parameters:
      - description: Akses Role
        in: body
        name: login
        required: true
        schema:
          $ref: '#/definitions/dto.SetAccessRole'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/middlewares.ResponseError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/middlewares.ResponseError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/middlewares.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/middlewares.ResponseError'
      summary: Update Akses Role
      tags:
      - Auth
...
schemes:
- http
swagger: "2.0"